\name{DeducerSurvival-package}
\alias{DeducerSurvival-package}
\alias{DeducerSurvival}
\alias{makeKMDialog}
\alias{.KMCheckFunction}
\alias{.KMRunFunction}
\alias{makeCompareDialog}
\alias{.CompareCheckFunction}
\alias{.CompareRunFunction}
\title{
  DeducerSurvivalDialog
}
\description{
  Adds Kaplan Meier Analysis GUI to Deducer
}
\usage{
makeKMDialog()
.KMCheckFunction(state)
.KMRunFunction(state)
makeCompareDialog()
.CompareCheckFunction(state)
.CompareRunFunction(state)
}

\arguments{
     \item{state}{ a list representing the current state of all of the widgets in the dialog}
 }

 \details{
 
 
 see http://www.deducer.org for details
 
 }
